LINE ITEM MODEL TEXT FILE REVISED
Line item model text file was as the same as generated with scaffold command.
This commit updates tests for line item model considering its attributes in the database.
Checks if they are not empty, if they are valid, if data matches and if updating.

test 'Time of the second addition of a line item registered must be different from first line item added'
might need some extra work as date and time are matching when they are not suppose to.